refactor: separate concerns into testable modules
